RAG_chatbot_confluence/requirements.backend.txt
fastapi==0.98.0
uvicorn[standard]==0.23.1
requests==2.31.0
azure-search-documents==11.7.0
azure-core==1.29.0
pydantic==1.10.11
